fix(ios-crash): bug with textselection causing crashes when the indices dont exist
