+++ app-swift/Fushigi/Utilities/Constants.swift
@@ -1,0 +1,18 @@
+//
+//  Constants.swift
+//  Fushigi
+//
+//  Created by Tahoe Schrader on 2025/08/19.
+//
+
+import SwiftUI
+
+private enum UIConstants {
+    static let contentMinHeight: CGFloat = 150
+    static let defaultSpacing: CGFloat = 20
+    static let sectionSpacing: CGFloat = 16
+    static let rowSpacing: CGFloat = 8
+    static let defaultPadding: CGFloat = 10
+    static let borderWidth: CGFloat = 1
+    static let fontSize: CGFloat = 18
+}
